Reorganize the schema files
