+++ packages/data-access-plugin/src/manager/schema/index.ts
@@ -2,9 +2,9 @@ import GraphQLJSON from 'graphql-type-json';
 import { GraphQLDateTime } from 'graphql-iso-date';
 import * as a from 'apollo-server-express';
 import * as d from '@terascope/data-access';
-import * as query from './query-schema';
-import * as mutation from './mutation-schema';
-import { ManagerContext } from './interfaces';
+import * as query from './query';
+import * as mutation from './mutation';
+import { ManagerContext } from '../interfaces';
 
 const resolvers: a.IResolvers<any, ManagerContext> = {
     Query: query.resolvers,
